refactor: moved item validation from server to gateway, fix tests
