--- gateway/src/main/java/ru/practicum/shareit/booking/dto/BookItemRequestDto.java
@@ -1,21 +1,0 @@
-package ru.practicum.shareit.booking.dto;
-
-import java.time.LocalDateTime;
-
-import javax.validation.constraints.Future;
-import javax.validation.constraints.FutureOrPresent;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-public class BookItemRequestDto {
-	private long itemId;
-	@FutureOrPresent
-	private LocalDateTime start;
-	@Future
-	private LocalDateTime end;
-}
